Oag 69 general refactor (#34)
* Made array syntax consistent

* Removed unused code en masse

* Distributed actions amongst sanely named controllers

* Split and thinned out controllers into services where appropriate

* Rename ActivityService to IATIService

* Rename IATIService to IATI

* Add documentation to services where appropriate

* Adjust test following refactor

* Revert accidental adjustment to fixture XML file

* Add code tweaks/fixes.
