+++ src/OagBundle/Controller/CoveController.php
@@ -2,12 +2,15 @@
 
 namespace OagBundle\Controller;
 
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use OagBundle\Entity\OagFile;
+use OagBundle\Service\IATI;
+use OagBundle\Service\Cove;
+use OagBundle\Service\OagFileService;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
-use OagBundle\Service\Cove;
-use OagBundle\Entity\OagFile;
-use OagBundle\Service\OagFileService;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * @Route("/cove")
@@ -16,46 +19,60 @@ use OagBundle\Service\OagFileService;
 class CoveController extends Controller {
 
     /**
-     * @Route("/{fileid}", requirements={"fileid": "\d+"})
-     * @Template
+     * Process on OagFile with CoVE.
+     *
+     * @Route("/{id}")
+     * @ParamConverter("file", class="OagBundle:OagFile")
      */
-    public function indexAction($fileid) {
-        $messages = [];
+    public function oagFileAction(Request $request, OagFile $file) {
         $cove = $this->get(Cove::class);
         $srvOagFile = $this->get(OagFileService::class);
+        $srvIATI = $this->get(IATI::class);
 
-        $repository = $this->getDoctrine()->getRepository(OagFile::class);
-        $oagfile = $repository->find($fileid);
-        if (!$oagfile) {
-            throw $this->createNotFoundException(sprintf('The document %d does not exist', $fileid));
-        }
-        $this->get('logger')->debug(sprintf('Processing %s using CoVE', $oagfile->getDocumentName()));
+        $this->get('logger')->debug(sprintf('Processing %s using CoVE', $file->getDocumentName()));
         // TODO - for bigger files we might need send as Uri
-        $path = $srvOagFile->getPath($oagfile);
-        $contents = file_get_contents($path);
+        $contents = $srvOagFile->getContents($file);
         $json = $cove->processString($contents);
 
-        $xml = $json['xml'];
-        $xmldir = $this->getParameter('oagxml_directory');
-        if (!is_dir($xmldir)) {
-            mkdir($xmldir, 0755, true);
-        }
-        $filename = $srvOagFile->getXMLFileName($oagfile);
-        $xmlfile = $xmldir . '/' . $srvOagFile->getXMLFileName($oagfile);
-        file_put_contents($xmlfile, $xml);
-
-        $err = $json['err'] ?? '';
+        $err = array_filter($json['err'] ?? array());
         $status = $json['status'] ?? '';
 
-        $pretty_json = json_encode($json, JSON_PRETTY_PRINT);
-        return array(
-            'messages' => $messages,
-            'path' => $oagfile->getDocumentName(),
-            'xml' => $xmlfile,
-            'err' => $err,
-            'status' => $status,
-            'id' => $oagfile->getId(),
-        );
+        // TODO Check status
+        foreach ($err as $line) {
+            $this->get('session')->getFlashBag()->add('error', $line);
+        }
+
+        $xml = $json['xml'];
+        if ($srvIATI->parseXML($xml)) {
+            $xmldir = $this->getParameter('oagxml_directory');
+            if (!is_dir($xmldir)) {
+                mkdir($xmldir, 0755, true);
+            }
+
+            $filename = $srvOagFile->getXMLFileName($file);
+            $xmlfile = $xmldir . '/' . $filename;
+            if (!file_put_contents($xmlfile, $xml)) {
+                $this->get('session')->getFlashBag()->add('error', 'Unable to create XML file.');
+                $this->get('logger')->debug(sprintf('Unable to create XML file: %s', $xmlfile));
+                return $this->redirectToRoute('oag_default_index');
+            }
+            // else
+            if ($this->getParameter('unlink_files')) {
+                unlink($path);
+            }
+
+            $file->setDocumentName($filename);
+            $file->setFileType(OagFile::OAGFILE_IATI_DOCUMENT);
+            $file->setMimeType('application/xml');
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($file);
+            $em->flush();
+            $this->get('session')->getFlashBag()->add('info', 'IATI File created/Updated\; ' . $xmlfile);
+        } else {
+            $this->get('session')->getFlashBag()->add('error', 'CoVE returned data that was not XML.');
+        }
+
+        return $this->redirectToRoute('oag_default_index');
     }
 
 }
