+++ src/OagBundle/Controller/DPortalController.php
@@ -19,7 +19,7 @@ class DPortalController extends Controller {
      * @Route("/{id}")
      * @ParamConverter("file", class="OagBundle:OagFile")
      */
-    public function indexAction(Request $request, OagFile $file) {
+    public function oagFileAction(Request $request, OagFile $file) {
         $portal = $this->get(DPortal::class);
 
         if ($portal->isAvailable()) {
