+++ src/OagBundle/Controller/DefaultController.php
@@ -8,7 +8,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use OagBundle\Service\Cove;
 use Symfony\Component\HttpFoundation\Request;
 use OagBundle\Entity\OagFile;
-use OagBundle\Service\ActivityService;
+use OagBundle\Service\IATI;
 use OagBundle\Service\OagFileService;
 use OagBundle\Form\OagFileType;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
@@ -63,29 +63,19 @@ class DefaultController extends Controller {
                 $em->persist($oagfile);
                 $em->flush();
 
-                return $this->redirect($this->generateUrl('oag_oagfile_source', ['id' => $oagfile->getId()]));
+                return $this->redirect($this->generateUrl('oag_cove_oagfile', array('id' => $oagfile->getId())));
             }
         }
 
-        $sourcedocs = $data = [
+        $sourcedocs = $data = array( 
             'iatidocs' => $iatidocs,
             'sourceDocs' => $sourceDocs,
             'enhancingDocs' => $enhancingDocs,
             'source_upload_form' => $sourceUploadForm->createView(),
-        ];
+        );
         return $data;
     }
-
-    /**
-     * Load oag files by file type.
-     */
-    private function loadOagFileByType($type) {
-        $repository = $this->getDoctrine()->getRepository(OagFile::class);
-        $oagfiles = $repository->findByFileType($type);
-
-        return $oagfiles;
-    }
-
+    
     /**
      * @Route("/delete/{ids}")
      */
@@ -110,65 +100,16 @@ class DefaultController extends Controller {
         return $this->redirectToRoute('oag_default_index');
     }
 
-    /**
-     * @Route("/upload")
-     * @Template
-     */
-    public function uploadAction(Request $request) {
 
-        $em = $this->getDoctrine()->getManager();
-        $oagfile = new OagFile();
-        $form = $this->createForm(OagFileType::class, $oagfile);
-        $form->add('submit', SubmitType::class, array(
-            'attr' => array('class' => 'submit'),
-        ));
-
-        // TODO Do something if the form is not valid
-        if ($request) {
-            $form->handleRequest($request);
-
-            // TODO Check for too big files.
-            if ($form->isSubmitted() && $form->isValid()) {
-                $file = $oagfile->getDocumentName();
-                $tmpFile = $oagfile->getDocumentName();
-                $oagfile->setMimeType(mime_content_type($tmpFile->getPathname()));
-
-                $filename = $file->getClientOriginalName();
-
-                $file->move(
-                    $this->getParameter('oagfiles_directory'), $filename
-                );
-
-                $oagfile->setDocumentName($filename);
-                $em->persist($oagfile);
-                $em->flush();
-
-                return $this->redirect($this->generateUrl('oag_default_index'));
-            }
-        }
-
-        return array(
-            'form' => $form->createView(),
-        );
-    }
 
     /**
-     * @Route("/activities")
-     * @Template
-     *
-     * TODO - link to a file in the database instead of using fixtures
+     * Load oag files by file type.
      */
-    public function activitiesAction() {
-        $srvActivity = $this->get(ActivityService::class);
-        $root = $srvActivity->parseXML($srvActivity->getFixtureData());
-        $simple = $srvActivity->summariseToArray($root);
+    private function loadOagFileByType($type) {
+        $repository = $this->getDoctrine()->getRepository(OagFile::class);
+        $oagfiles = $repository->findByFileType($type);
 
-        $files = array();
-        $files['non-fixture file name goes here'] = $simple;
-
-        return array(
-            'files' => $files
-        );
+        return $oagfiles;
     }
 
 }
