+++ src/OagBundle/Resources/views/Activity/enhance.html.twig
@@ -41,7 +41,7 @@
     <p>Any smart ideas about versioning welcome</p>
 
     <p>
-        <a href='{{ path('oag_oagfile_classifyxml', { 'id': id  }) }}'>Classify</a>
+        <a href='{{ path('oag_classify_oagfile', { 'id': id  }) }}'>Classify</a>
     </p>
 
     <p>
