--- src/OagBundle/Resources/views/Activity/index.html.twig
@@ -1,7 +1,0 @@
-{% extends "::base.html.twig" %}
-
-{% block title %}OagBundle:Activity:index{% endblock %}
-
-{% block body %}
-<h1>Welcome to the Activity:index page</h1>
-{% endblock %}
