--- src/OagBundle/Resources/views/Activity/merge.html.twig
@@ -1,7 +1,0 @@
-{% extends "::base.html.twig" %}
-
-{% block title %}OagBundle:Activity:merge{% endblock %}
-
-{% block body %}
-<h1>Welcome to the Activity:merge page</h1>
-{% endblock %}
