--- src/OagBundle/Resources/views/Classify/tags.html.twig
@@ -1,14 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block body %}
-    <h2>Add documents to enhance content</h2>
-    <div style='margin-bottom: 20px; display: inline-block'>
-        {% form_theme enhancementForm 'OagBundle::form/listEnhancementDocs.html.twig' %}
-        {{ form(enhancementForm) }}
-    </div>
-    <h2>Activities</h2>
-    <div style='margin-bottom: 20px; display: inline-block'>
-        {% form_theme tagsForm 'OagBundle::form/mergeActivity.html.twig' %}
-        {{ form(tagsForm) }}
-    </div>
-{% endblock %}
