--- src/OagBundle/Resources/views/Cove/index.html.twig
@@ -1,19 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block title %}CoVE{% endblock %}
-
-{% block h2 %}CoVE{% endblock %}
-
-{% block body %}
-
-    {% for message in messages %}
-        <div>{{ message }}</div>
-    {% endfor %}
-
-    <ul>
-        <li>Process status: {{ status }}</li>
-        <li>StdErr: <ul>{% for line in err %}<li>{{ line }}<br /></li>{% endfor %}</ul></li>
-        <li><a href="{{ asset('xml')}}/oagfiles/{{ path }}">XML</a></li>
-    </ul>
-
-{% endblock %}
