--- src/OagBundle/Resources/views/Default/activities.html.twig
@@ -1,42 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block stylesheets %}
-    <style>
-        #tags li, #locations li {
-            display: inline;
-            list-style-type: none;
-            background-color: lightgrey;
-            margin-right: 1em;
-        }
-    </style>
-{% endblock %}
-
-{% block body %}
-    <ul id='files'>
-        {% for fileName, activities in files %}
-            <li>
-                <h1>{{ fileName }}</h1>
-                <ul id='activities'>
-                    {% for activity in activities %}
-                        <li>
-                            <h2>{{ activity.name }} <em>({{ activity.id }})</em></h2>
-                            <h3>Tags</h3>
-                            <ul id='tags'>
-                                {% for tag in activity.tags %}
-                                    <li>{{ tag.description }}</li>
-                                    {% endfor %}
-                            </ul>
-                            <h3>Locations</h3>
-                            <ul id='locations'>
-                                {% for location in activity.locations %}
-                                    <li>{{ location.description }}</li>
-                                    {% endfor %}
-                            </ul>
-                        </li>
-                        <hr>
-                    {% endfor %}
-                </ul>
-            </li>
-        {% endfor %}
-    </ul>
-{% endblock %}
