+++ src/OagBundle/Resources/views/Default/index.html.twig
@@ -17,7 +17,7 @@
     <ul>
         {% for iatidoc in iatidocs %}
             <li>
-                <a href="{{ path('oag_dportal_index', { 'id': iatidoc.id }) }}" target="dportal">[Visualise]</a>
+                <a href="{{ path('oag_dportal_oagfile', { 'id': iatidoc.id }) }}" target="dportal">[Visualise]</a>
                 {{ include('OagBundle::Default/iatiFile.html.twig', { 'iatiDoc': iatidoc }) }}
                 <ul>
                     {% for enhancingDoc in iatidoc.enhancingDocuments %}
