+++ src/OagBundle/Resources/views/Default/sourceFile.html.twig
@@ -1,1 +1,1 @@
-<a href="{{ path('oag_oagfile_source', { 'id': sourceDoc.id }) }}">{{ sourceDoc.documentName }}</a>
+<a href="{{ path('oag_cove_oagfile', { 'id': sourceDoc.id }) }}">{{ sourceDoc.documentName }}</a>
