--- src/OagBundle/Resources/views/Default/upload.html.twig
@@ -1,22 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block title %}Upload{% endblock %}
-
-{% block h2 %}Upload{% endblock %}
-
-{% block body %}
-
-    <h2>Upload new file</h2>
-
-    <div>
-        {% if form is defined and form is not null %}
-            {{ form_start(form) }}
-            {{ form_widget(form) }}
-            {{ form_end(form) }}
-
-        {% else %}
-            NO FORM
-        {% endif %}
-    </div>
-
-{% endblock %}
