--- src/OagBundle/Resources/views/Geocode/locations.html.twig
@@ -1,14 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block stylesheets %}
-    <style>
-        td, th, table {
-            border: 2px solid black;
-        }
-    </style>
-{% endblock %}
-
-{% block body %}
-    {% form_theme form 'OagBundle::form/mergeActivity.html.twig' %}
-    {{ form(form) }}
-{% endblock %}
