--- src/OagBundle/Resources/views/Geocode/mergeLocations.html.twig
@@ -1,13 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block body %}
-
-    {{ form_start(form) }}
-    {{ form_widget(form) }}
-    {{ form_end(form) }}
-
-    {% if processed is defined %}
-        <pre>{{ processed }}</pre>
-    {% endif %}
-
-{% endblock %}
