+++ src/OagBundle/Resources/views/Geocode/oagFile.html.twig
@@ -8,11 +8,10 @@
     <h2>GEOCODER DOES NOT WORK</h2>
     <h3>Returning fixtures</h3>
     {% if geolocations|length > 0 %}
-        {{ include('OagBundle::OagFile/geolocations_table.html.twig', { 'geolocations': geolocations }) }}
+        {{ include('OagBundle::Geocode/geolocations_table.html.twig', { 'geolocations': geolocations }) }}
     {% else %}
         <h2>No Geolocations associated</h2>
     {% endif %}
-    <pre>{{ json }}</pre>
     <ul>
         <li>Store results alongside source doc</li>
         <li>Do we store JSON or start a folksonomy of geocoded location referrenced by id?</li>
