--- src/OagBundle/Resources/views/OagFile/classifyText.html.twig
@@ -1,14 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block title %}Classify results: {{ name }}{% endblock %}
-
-{% block h2 %}{{ name }}{% endblock %}
-
-{% block body %}
-    {% if tags|length > 0 %}
-        <h2>Tags</h2>
-        {{ include('OagBundle::OagFile/suggested_tags_table.html.twig', { 'tags': tags }) }}
-    {% else %}
-        <h2>No tags associated</h2>
-    {% endif %}
-{% endblock %}
