+++ src/OagBundle/Resources/views/OagFile/enhancement.html.twig
@@ -8,19 +8,19 @@
     <h2>Mime type: {{ mimetype }}</h2>
 
     {% if tags|length > 0 %}
-        {{ include('OagBundle::OagFile/suggested_tags_table.html.twig', { 'tags': tags }) }}
+        {{ include('OagBundle::Classify/suggested_tags_table.html.twig', { 'tags': tags }) }}
     {% else %}
         <h2>No tags associated</h2>
     {% endif %}
 
     {% if geolocations|length > 0 %}
-        {{ include('OagBundle::OagFile/geolocations_table.html.twig', { 'geolocations': geolocations }) }}
+        {{ include('OagBundle::Geocode/geolocations_table.html.twig', { 'geolocations': geolocations }) }}
     {% else %}
         <h2>No Geolocations associated</h2>
     {% endif %}
     <div>
-        <a href='{{ path('oag_oagfile_classifytext', { 'id': id }) }}'>{% if tags|length > 0 %}Re-{% endif %}Classify</a>
+        <a href='{{ path('oag_classify_oagfile', { 'id': id }) }}'>{% if tags|length > 0 %}Re-{% endif %}Classify</a>
         |
-        <a href='{{ path('oag_oagfile_geocode', { 'id': id }) }}'>{% if tags|length > 0 %}Re-{% endif %}Geocode</a>
+        <a href='{{ path('oag_geocode_oagfile', { 'id': id }) }}'>{% if tags|length > 0 %}Re-{% endif %}Geocode</a>
     </div>
 {% endblock %}
