+++ src/OagBundle/Resources/views/OagFile/iati.html.twig
@@ -11,7 +11,7 @@
     |
     <a class='view-xml-link' href="{{ path('oag_oagfile_raw', { 'id': id }) }}">Raw XML</a>
     |
-    <a class='view-xml-link' href="{{ path('oag_oagfile_classifyxml', { 'id': id }) }}">Re-Classify</a>
+    <a class='view-xml-link' href="{{ path('oag_classify_oagfile', { 'id': id }) }}">Re-Classify</a>
 {% endblock %}
 
 {% block body %}
@@ -33,8 +33,8 @@
                     {% for doc in enhancingDocuments %}
                         <li><span class='docname'>{{ doc.documentName }}</span>
                             <a href="{{ path('oag_oagfile_enhancement', { 'id': doc.id }) }}">View</a> |
-                            <a href='{{ path('oag_oagfile_classifytext', { 'id': doc.id }) }}'>{% if doc.suggestedtags|length > 0 %}Re-{% endif %}Classify</a> |
-                            <a href='{{ path('oag_oagfile_geocode', { 'id': doc.id }) }}'>{% if doc.geolocations|length > 0 %}Re-{% endif %}Geocode</a>
+                            <a href='{{ path('oag_classify_oagfile', { 'id': doc.id }) }}'>{% if doc.suggestedtags|length > 0 %}Re-{% endif %}Classify</a> |
+                            <a href='{{ path('oag_geocode_oagfile', { 'id': doc.id }) }}'>{% if doc.geolocations|length > 0 %}Re-{% endif %}Geocode</a>
                         {% endfor %}
                 </ul>
         {% else %}
