--- src/OagBundle/Resources/views/Tag/index.html.twig
@@ -1,24 +1,0 @@
-{% extends 'OagBundle::Default/index.html.twig' %}
-
-{% block body %}
-
-    <table class="pure-table pure-table-bordered">
-        <thead>
-            <tr>
-                <th>AGROVOC Code</th>
-                <th>Descroption</th>
-                <th>Confidence</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for row in tags %}
-                <tr>
-                    <td>{{ row.getCode.getCode }}</td>
-                    <td>{{ row.getCode.getDescription }}</td>
-                    <td>{{ row.getConfidence }}</td>
-                </tr>
-            {% endfor %}
-        </tbody>
-    </table>
-
-{% endblock %}
