+++ src/OagBundle/Service/AbstractOagService.php
@@ -4,8 +4,18 @@ namespace OagBundle\Service;
 
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
+/**
+ * Represents a service that may be enacted on the contents of an OagFIle but
+ * that is not guaranteed to always be available.
+ */
 abstract class AbstractOagService extends AbstractService {
 
+    /**
+     * Get the URI endpoint of the external service.
+     *
+     * @param $key the key in the YAML config file of the URI
+     * @return string the URI as a string
+     */
     public function getUri($key = 'uri') {
         $oag = $this->getContainer()->getParameter('oag');
 
@@ -15,6 +25,12 @@ abstract class AbstractOagService extends AbstractService {
         return $uri;
     }
 
+    /**
+     * Work out whether the external service is accessible or whether we have
+     * to fall back to fixture data.
+     *
+     * @return boolean TRUE if available, else FALSE
+     */
     public function isAvailable() {
         $name = $this->getName();
         $cmd = sprintf('docker images openagdata/%s |wc -l', $name);
