+++ src/OagBundle/Service/Geocoder.php
@@ -2,6 +2,7 @@
 
 namespace OagBundle\Service;
 
+use OagBundle\Entity\Geolocation;
 use OagBundle\Entity\OagFile;
 
 class Geocoder extends AbstractAutoService {
@@ -37,18 +38,87 @@ class Geocoder extends AbstractAutoService {
         return $contents;
     }
 
-    public function extractLocations($json) {
-        $locations = array();
-        foreach ($json as $projectObject) {
-            $id = $projectObject['project_id'];
-            $projLocations = $projectObject['locations'];
-            $locations[$id] = $projLocations;
-        }
-        return $locations;
-    }
-
     public function processOagFile(OagFile $file) {
         return $this->getFixtureData();
+    }
+
+    /**
+     * Process an OagFile and persist the results as Geolocation objects,
+     * attached to the OagFile.
+     *
+     * @param OagFile $file the file to process
+     */
+    public function geocodeOagFile(OagFile $file) {
+        $json = $this->processOagFile($file);
+        $results = json_decode($json, true);
+
+        $file->clearGeolocations();
+        $em = $this->getContainer()->get('doctrine')->getManager();
+        $geolocationrepo = $this->getContainer()->get('doctrine')->getRepository(Geolocation::class);
+
+        foreach ($results as $activity) {
+            $iatiActivityId = $activity['project_id'] ?? null;
+            $locations = $activity['locations'];
+            foreach ($locations as $location) {
+                $locationId = $location['id'];
+                $vocabId = '99'; // TODO get a valif vocab id
+                // Does this location already exist for this IATI ID?
+                $geolocation = $geolocationrepo->findOneBy(
+                    array(
+                        'iatiActivityId' => $iatiActivityId,
+                        'geolocationId' => $locationId,
+                        'vocabId' => $vocabId,
+                    )
+                );
+
+                if (!$geolocation) {
+                    $geolocation = new Geolocation();
+                    $geolocation->setIatiActivityId($iatiActivityId);
+                    $geolocation->setGeolocationId($locationId);
+                    $geolocation->setVocabId('99'); // TODO get a valif vocab id
+                }
+                $geolocation->setName($location['name']);
+                $geolocation->setAdminCode1Code($location['admin1']['code']);
+                $geolocation->setAdminCode1Name($location['admin1']['name']);
+                $geolocation->setAdminCode2Code($location['admin2']['code']);
+                $geolocation->setAdminCode2Name($location['admin2']['name']);
+                $geolocation->setLatitude($location['geometry']['coordinates'][0]);
+                $geolocation->setLongitude($location['geometry']['coordinates'][1]);
+                $geolocation->setExactness($location['exactness']['code']);
+                $geolocation->setClass($location['locationClass']['code']);
+                $geolocation->setDescription($location['locationClass']['description']);
+                $em->persist($geolocation);
+
+                if (!$file->hasGeolocation($geolocation)) {
+                    $file->addGeolocation($geolocation);
+                }
+            }
+        }
+        $em->persist($file);
+        $em->flush();
+    }
+
+    /**
+     * Flatten a geolocation entity instance to an associative array summarising it.
+     *
+     * @param Geolocation $location
+     * @return array see function for clearest view of keys and values
+     */
+    public function locationToArray(Geolocation $location) {
+        $data = array();
+        $data['vocab_id'] = $location->getVocabId();
+        $data['geolocation_id'] = $location->getGeolocationId();
+        $data['name'] = $location->getName();
+        $data['admin_code_1_code'] = $location->getAdminCode1Code();
+        $data['admin_code_1_name'] = $location->getAdminCode1Name();
+        $data['admin_code_2_code'] = $location->getAdminCode2Code();
+        $data['admin_code_2_name'] = $location->getAdminCode2Name();
+        $data['latitude'] = $location->getLatitude();
+        $data['longitude'] = $location->getLongitude();
+        $data['exactness'] = $location->getExactness();
+        $data['class'] = $location->getClass();
+        $data['description'] = $location->getDescription();
+        return $data;
     }
 
 }
